# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Appointment Scheduling API",
        "description": "This is a simple API for managing appointments.",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/appointments/available": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get available appointment times",
                "description": "Retrieve available appointment times for new patients, consultations, and follow-ups",
                "operationId": "f41e4454c22d6065954f152dfc0fbeef",
                "parameters": [
                    {
                        "name": "next_number_of_days",
                        "in": "query",
                        "description": "Number of days in the future to retrieve available appointment slots",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 7
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "new_patient": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "date": {
                                                        "type": "string",
                                                        "example": "2025-04-04"
                                                    },
                                                    "doctor": {
                                                        "type": "string",
                                                        "example": "Hermione Granger"
                                                    },
                                                    "doctor_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "available_start_time": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string",
                                                            "example": "12:00"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "consultation": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "date": {
                                                        "type": "string",
                                                        "example": "2025-04-04"
                                                    },
                                                    "doctor": {
                                                        "type": "string",
                                                        "example": "Ron Weasley"
                                                    },
                                                    "doctor_id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "available_start_time": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string",
                                                            "example": "09:00"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "follow_up": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "date": {
                                                        "type": "string",
                                                        "example": "2025-04-05"
                                                    },
                                                    "doctor": {
                                                        "type": "string",
                                                        "example": "Draco Malfoy"
                                                    },
                                                    "doctor_id": {
                                                        "type": "integer",
                                                        "example": 4
                                                    },
                                                    "available_start_time": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string",
                                                            "example": "13:00"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/doctors/{id}": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Get a specific doctor by ID",
                "description": "Retrieve a specific doctor's details, including their name, email, and schedules",
                "operationId": "f7d27b7312d2bcfb24da4bf9b7109e21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the doctor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with doctor details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Harry Potter"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "harry.potter@example.com"
                                        },
                                        "schedules": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "day_of_week": {
                                                        "type": "string",
                                                        "example": "Sunday"
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "format": "time",
                                                        "example": "09:00:00"
                                                    },
                                                    "end_time": {
                                                        "type": "string",
                                                        "format": "time",
                                                        "example": "12:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Doctor not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Doctor not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}